Fix orders service endpoints
