+++ src/Services/Ordering/Microstore.Service.OrderingApplication/Orders/EventHandlers/Domain/OrderUpdatedEventHandler.cs
@@ -1,9 +1,9 @@
 ﻿namespace Microstore.Service.OrderingApplication.Orders.EventHandlers.Domain;
-public class OrderCreatedEventHandler
-    (ILogger<OrderCreatedEventHandler> logger)
-    : INotificationHandler<OrderCreatedEvent>
+public class OrderUpdatedEventHandler
+    (ILogger<OrderUpdatedEventHandler> logger)
+    : INotificationHandler<OrderUpdatedEvent>
 {
-    public Task Handle(OrderCreatedEvent notification, CancellationToken cancellationToken)
+    public Task Handle(OrderUpdatedEvent notification, CancellationToken cancellationToken)
     {
         logger.LogInformation("Domain Event handled: {DomainEvent}", notification.GetType().Name);
         return Task.CompletedTask;
